Add Mattermost notifications functionality
Implemented the functionality for creating Mattermost notifications for reminders. This includes updates to the submit form for notifications, necessary queries, adjustments to the database schema for notifications and UI updates in the client-notifications page. This change was needed to provide support for Mattermost notifications, expanding the range of notification channels our application can cater to.