Remove user signup process and simplifying login
Deleted the user signup function in server handlers and corresponding URL routes, and also removed related HTML templates. This is because we are transitioning to a single-user system that utilizes a pre-shared key rather than user-specific email and password information. As part of these changes, the location of the login page has been updated, and the database queries and form validation for login process have been simplified. These changes will streamline the login process and help improve application performance.