Enhanced user interactivity with reminder option and delete function
The update comprises of two key enhancements aimed to improve the user experience. Firstly, an update was made allowing users to select specific days to execute reminders in the "mattermost-form.tmpl". A small description was added to guide users to select one or multiple days. Secondly, Delete functionality was added that gives users the ability to remove specific clients and notifications. HMTX was used to handle the interactions with server and swapping to enable a smooth transition for users.

Repository handling functions were implemented for notification and client deletion. Now, when the delete button is clicked, a confirmation message is displayed and html content is dynamically updated without page refresh. Styling changes were added to make this transition smoother. Unit for repository and client deletion were implemented with database queries providing essential functionality.

The changes were motivated by the goal of enhancing the user experience of the application and providing greater customizability and interactivity when setting reminders and managing clients and notifications.